refactor: must helper only used within locations so move it there
