+++ internal/config/locations/locations.go
@@ -5,7 +5,6 @@ import (
 	"path/filepath"
 
 	"github.com/aripalo/vegas-credentials/internal/config"
-	"github.com/aripalo/vegas-credentials/internal/utils"
 
 	"github.com/adrg/xdg"
 )
@@ -19,12 +18,12 @@ var CacheDir = EnsureWithinDir(xdg.CacheHome, config.AppName)
 var StateDir = EnsureWithinDir(xdg.StateHome, config.AppName)
 
 // The directory where vegas-credentials executable is located.
-var ExecDir = filepath.Dir(utils.Must(os.Executable))
+var ExecDir = filepath.Dir(must(os.Executable))
 
 // AWS config file location. Usually in $HOME/.aws/config or %USERPROFILE%/.aws/config
 // unless user has set $AWS_CONFIG_FILE environment variable.
 // https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-files.html#cli-configure-files-where
-var AwsConfig = utils.Must(getAwsConfigFile)
+var AwsConfig = must(getAwsConfigFile)
 
 // Location of the Yubikey Manager CLI (ykman) executable.
 // Empty if not available.
